270 finish page with zrange
ZRANGE turns out to be part of the set discussion on p. 270,
I stopped at the subsection header yesterday. I should read
ahead through the material before starting to implement what's
on the page.

This commit also show some changes to specs which worked
yesterday and failed this morning. Very odd. I did have to
pull a new docker image, and I believe the specs were written
on a different computer running a different instance of redis.
The instance which was used for these specs is version 5.0.8.
I'll have to keep an eye on this when I move back to the other
computer tomorrow.